Add better exception handling

+++ src/main/java/org/example/cloudstorage/service/FileService.java
@@ -5,6 +5,7 @@ import org.example.cloudstorage.domain.dto.FileDetailsDto;
 import org.example.cloudstorage.domain.entity.File;
 import org.example.cloudstorage.domain.entity.User;
 import org.example.cloudstorage.domain.entity.UserStorage;
+import org.example.cloudstorage.exception.*;
 import org.example.cloudstorage.mapper.FileDetailsDtoMapper;
 import org.example.cloudstorage.mapper.PublicUserDtoMapper;
 import org.example.cloudstorage.repository.FileRepository;
@@ -43,27 +44,27 @@ public class FileService {
         this.fileDetailsDtoMapper = fileDetailsDtoMapper;
     }
 
-    public FileDetailsDto saveFile(MultipartFile mpFile, String userName) throws IOException {
-        User owner = userRepository.findUserByUserName(userName);
-        File file = new File(owner, mpFile.getOriginalFilename(), mpFile.getContentType(), mpFile.getSize(), saveToFileSystem(mpFile, userName));
+    public FileDetailsDto saveFile(MultipartFile mpFile, String ownerName) {
+        User owner = userRepository.findUserByUserName(ownerName);
+        File file = new File(owner, mpFile.getOriginalFilename(), mpFile.getContentType(), mpFile.getSize(), saveToFileSystem(mpFile, ownerName));
         File dbFile = fileRepository.save(file);
         userStorageRepository.save(new UserStorage(owner, dbFile));
         return fileDetailsDtoMapper.apply(dbFile, userRepository.findShareUsersByFileId(dbFile.getId()));
     }
 
-    public FileDataDto readFile(Long id, String userName) throws IOException {
-        if (userName.equals(fileRepository.findById(id).get().getOwner().getName())) {
-            File file = fileRepository.findById(id).get();
+    public FileDataDto readFile(Long fileId, String ownerName) {
+        File file = fileRepository.findById(fileId).orElseThrow(FileNotFoundException::new);
+        if (ownerName.equals(file.getOwner().getName())) {
             byte[] bytes = readFromFileSystem(file.getPath());
             return new FileDataDto(file.getName(), bytes);
         } else {
-            throw new RuntimeException("Wrong authentication!");
+            throw new UserNotAuthorizedException();
         }
     }
 
     public Page<FileDetailsDto> findUserFiles(String userName, String ownerName, String fileType, Pageable pageable) {
         if (!File.hasValidSortField(pageable.getSort())) {
-            throw new RuntimeException("Invalid sort field!");
+            throw new InvalidSortParameterException();
         }
         List<FileDetailsDto> mappedFiles = new ArrayList<>();
         Page<File> page = fileRepository.findFilesByUserName(userName, ownerName, fileType, pageable);
@@ -71,51 +72,67 @@ public class FileService {
         return page.map(file -> fileDetailsDtoMapper.apply(file, userRepository.findShareUsersByFileId(file.getId())));
     }
 
-    public void deleteFile(Long fileId, String userName) {
-        if (userName.equals(fileRepository.findById(fileId).get().getOwner().getName())) {
-            File file = fileRepository.findById(fileId).get();
+    public void deleteFile(Long fileId, String ownerName) {
+        File file = fileRepository.findById(fileId).orElseThrow(FileNotFoundException::new);
+        if (ownerName.equals(file.getOwner().getName())) {
             if (deleteFileFromFileSystem(file.getPath())) {
                 fileRepository.deleteById(fileId);
             }
         } else {
-            throw new RuntimeException("Wrong authentication!");
+            throw new UserNotAuthorizedException();
         }
     }
 
     public FileDetailsDto shareWithUser(Long fileId, String userName, String ownerName) {
-        File file = fileRepository.findById(fileId).get();
+        File file = fileRepository.findById(fileId).orElseThrow(FileNotFoundException::new);
         if (ownerName.equals(file.getOwner().getName())) {
             User user = userRepository.findUserByUserName(userName);
+            if (user == null) {
+                throw new UserNotFoundException();
+            }
             userStorageRepository.save(new UserStorage(user, file));
             return fileDetailsDtoMapper.apply(file, userRepository.findShareUsersByFileId(fileId));
         } else {
-            throw new RuntimeException("Wrong authentication!");
+            throw new UserNotAuthorizedException();
         }
     }
 
     public FileDetailsDto unshareWithUser(Long fileId, String userName, String ownerName) {
-        File file = fileRepository.findById(fileId).get();
+        File file = fileRepository.findById(fileId).orElseThrow(FileNotFoundException::new);
         if (ownerName.equals(file.getOwner().getName())) {
             User user = userRepository.findUserByUserName(userName);
+            if (user == null) {
+                throw new UserNotFoundException();
+            }
             userStorageRepository.delete(new UserStorage(user, file));
             return fileDetailsDtoMapper.apply(file, userRepository.findShareUsersByFileId(fileId));
         } else {
-            throw new RuntimeException("Wrong authentication!");
+            throw new UserNotAuthorizedException();
         }
     }
 
-    private String saveToFileSystem(MultipartFile mpFile, String userName) throws IOException {
+    private String saveToFileSystem(MultipartFile mpFile, String userName) {
         String parent = "%s/%s".formatted(storageLocation, userName);
         String path = "%s/%s".formatted(parent, mpFile.getOriginalFilename());
-        Files.createDirectories(Path.of(parent));
-        FileOutputStream outputStream = new FileOutputStream(path);
-        outputStream.write(mpFile.getBytes());
+        try {
+            Files.createDirectories(Path.of(parent));
+        } catch (IOException e) {
+            throw new FileNotWrittenException("Could not create the user directory.");
+        }
+        try (FileOutputStream outputStream = new FileOutputStream(path)) {
+            outputStream.write(mpFile.getBytes());
+        } catch (IOException e) {
+            throw new FileNotWrittenException();
+        }
         return path;
     }
 
-    private byte[] readFromFileSystem(String path) throws IOException {
-        FileInputStream inputStream = new FileInputStream(path);
-        return inputStream.readAllBytes();
+    private byte[] readFromFileSystem(String path) {
+        try (FileInputStream inputStream = new FileInputStream(path)) {
+            return inputStream.readAllBytes();
+        } catch (IOException e) {
+            throw new FileNotReadException();
+        }
     }
 
     private boolean deleteFileFromFileSystem(String path) {
